Add UCD to DataRequest.R
Specifically UCD_ICD_2020 - 1999.txt files as a non-provisional set by year. They are aggregated to show only the total for all months in the year. The same cannot be said for months where delays in the input and collection of the data for individuals is delayed, sometimes substantially. I suspect this is due to delays that result in a bunch of points being posted at once even though they may have happened months ago. At the CDC rather than enter these in, they ignore the month value and assign it to the default year. This default year, however, also offers room for error. For example, what happens when deaths occur in one year but changes are made in subsequent years? Provisional data should shed some light on this.