Add Activity with Time Crunch
Updated mortality csv by subsetting and sampling the data used in activity for all-staff meeting with minor changes to the time folder. Gave a clearer challenge and started to define a solution. The time_crunch script is only meant to reassess and test calculations of time in different ways than simply "how long did [x] project take." We have categories, use them! And determine if they would be helpful for later